requirements.txt
fastapi
uvicorn
httpx
streamlit
google-genai
python-dotenv
requests
